+++ Cargo.toml
@@ -50,6 +50,7 @@ eventsource-stream = "0.2"
 futures-util = "0.3"
 hex = "0.4"
 http = "1"
+http-body-util = "0.1"
 libc = "0.2"
 lancedb = "0.29"
 notify = "7"
@@ -65,6 +66,7 @@ serde_json = "1"
 serial_test = "3"
 sha2 = "0.10"
 silent = "2"
+subtle = "2"
 tantivy = "0.22"
 tauri = "2"
 tauri-build = "2"
