+++ crates/tiangong-core/src/scheduler/mod.rs
@@ -1,2 +1,3 @@
 pub use tiangong_scheduler::model;
 pub use tiangong_scheduler::store;
+pub use tiangong_scheduler::webhook;
