+++ crates/tiangong-scheduler/src/lib.rs
@@ -1,2 +1,7 @@
 pub mod model;
 pub mod store;
+pub mod webhook;
+
+pub(crate) const fn default_true() -> bool {
+    true
+}
