+++ crates/tiangong-server/Cargo.toml
@@ -21,7 +21,9 @@ scru128 = { workspace = true }
 chrono = { workspace = true }
 hmac = "0.12"
 hex = { workspace = true }
+http-body-util = { workspace = true }
 sha2 = { workspace = true }
+subtle = { workspace = true }
 reqwest = { workspace = true, default-features = false, features = ["rustls-tls"] }
 dirs = "6"
 libc = { workspace = true }
